Define dependencies for admin bundle
